deleted some codes that are not necessary & changed detection model(.lite)
