--- app/src/main/assets/retrained_labels.txt
@@ -1,5 +1,0 @@
-kakaoneo
-kakaoryan
-shoes
-spoon
-toothbrush
