+++ app/src/main/assets/retrained_labels_ccp.txt
@@ -1,0 +1,7 @@
+banana milk
+coke can
+fanta grape
+fanta oragne
+monster energy drink
+pepsi can
+sprite can
