--- app/src/main/java/com/example/android/alarmapp/DetailAlarmActivity.java
@@ -1,199 +1,0 @@
-package com.example.android.alarmapp;
-
-import android.graphics.Bitmap;
-import android.os.AsyncTask;
-import android.support.design.widget.FloatingActionButton;
-import android.support.v4.app.FragmentManager;
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-import android.text.method.ScrollingMovementMethod;
-import android.util.Log;
-import android.view.View;
-import android.widget.Button;
-import android.widget.TextView;
-
-import com.example.android.alarmapp.data.Alarm;
-import com.example.android.alarmapp.tflite.Classifier;
-import com.example.android.alarmapp.tflite.TensorFlowImageClassifier;
-import com.example.android.alarmapp.utils.AlarmActionUtils;
-import com.example.android.alarmapp.utils.LocationUtils;
-import com.wonderkiln.camerakit.CameraKitError;
-import com.wonderkiln.camerakit.CameraKitEvent;
-import com.wonderkiln.camerakit.CameraKitEventListener;
-import com.wonderkiln.camerakit.CameraKitImage;
-import com.wonderkiln.camerakit.CameraKitVideo;
-import com.wonderkiln.camerakit.CameraView;
-
-import java.util.List;
-import java.util.concurrent.Executor;
-import java.util.concurrent.Executors;
-
-import io.realm.Realm;
-
-/*
-*
-*  알람시 표시되는 화면
-*
-*  시간 + 메모 + 위치정보 + 알람 중지 버튼
-*
-*
-*/
-
-public class DetailAlarmActivity extends AppCompatActivity {
-
-    private static final String TAG ="DetailAlarmActivity";
-    private Realm mRealm;
-
-//    private static final String MODEL_PATH = "mobilenet_quant_v1_224.tflite";
-//    private static final String LABEL_PATH = "labels.txt";
-
-    private static final String MODEL_PATH = "optimized_graph.lite";
-    private static final String LABEL_PATH = "retrained_labels.txt";
-    private static final int INPUT_SIZE = 224;
-
-    private Classifier classifier;
-
-    private Executor executor = Executors.newSingleThreadExecutor();
-    private TextView textViewResult;
-    private FloatingActionButton btnDetectObject, btnToggleCamera;
-    private CameraView cameraView;
-
-    Button mTestButton;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.test_layout);
-        mRealm=Realm.getDefaultInstance();
-
-        showMissionDialog();
-
-        cameraView = findViewById(R.id.cameraView);
-        textViewResult = findViewById(R.id.textViewResult);
-        textViewResult.setMovementMethod(new ScrollingMovementMethod());
-
-        btnToggleCamera = findViewById(R.id.btnToggleCamera);
-        btnDetectObject = findViewById(R.id.btnDetectObject);
-
-        mTestButton = findViewById(R.id.btn_test_cancel);
-        mTestButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                AlarmActionUtils.stopVibrate();
-                AlarmActionUtils.stopMusic();
-            }
-        });
-
-        cameraView.addCameraKitListener(new CameraKitEventListener() {
-            @Override
-            public void onEvent(CameraKitEvent cameraKitEvent) {
-
-            }
-
-            @Override
-            public void onError(CameraKitError cameraKitError) {
-
-            }
-
-            @Override
-            public void onImage(CameraKitImage cameraKitImage) {
-                Bitmap bitmap = cameraKitImage.getBitmap();
-                bitmap = Bitmap.createScaledBitmap(bitmap, INPUT_SIZE, INPUT_SIZE, false);
-                final List<Classifier.Recognition> results = classifier.recognizeImage(bitmap);
-
-                textViewResult.setText(results.toString());
-
-                String topResult = results.get(0).getTitle(); // highest precision result(label)
-
-                if(topResult.equals("kakaoryan")) { // todo : random label
-                    Log.d(TAG, "onImage: alarm stop!");
-                    AlarmActionUtils.stopVibrate();
-                    AlarmActionUtils.stopMusic();
-                }
-            }
-
-            @Override
-            public void onVideo(CameraKitVideo cameraKitVideo) {
-
-            }
-        });
-
-        btnToggleCamera.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                cameraView.toggleFacing();
-            }
-        });
-
-        btnDetectObject.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                cameraView.captureImage();
-            }
-        });
-
-        initTensorFlowAndLoadModel();
-
-    }
-
-    @Override
-    protected void onDestroy() {
-        super.onDestroy();
-        AlarmActionUtils.stopVibrate();
-        AlarmActionUtils.stopMusic();
-        mRealm.close();
-
-        executor.execute(new Runnable() {
-            @Override
-            public void run() {
-                classifier.close();
-            }
-        });
-    }
-
-
-    @Override
-    protected void onResume() {
-        super.onResume();
-        cameraView.start();
-    }
-
-    @Override
-    protected void onPause() {
-        cameraView.stop();
-        super.onPause();
-    }
-
-    private void initTensorFlowAndLoadModel() {
-        executor.execute(new Runnable() {
-            @Override
-            public void run() {
-                try {
-                    classifier = TensorFlowImageClassifier.create(
-                            getAssets(),
-                            MODEL_PATH,
-                            LABEL_PATH,
-                            INPUT_SIZE);
-                    makeButtonVisible();
-                } catch (final Exception e) {
-                    throw new RuntimeException("Error initializing TensorFlow!", e);
-                }
-            }
-        });
-    }
-
-    private void showMissionDialog() {
-        FragmentManager fm = getSupportFragmentManager();
-        MissionFragment mMissionFragment = new MissionFragment();
-        mMissionFragment.show(fm, "mission dialog");
-    }
-
-    private void makeButtonVisible() {
-        runOnUiThread(new Runnable() {
-            @Override
-            public void run() {
-                btnDetectObject.setVisibility(View.VISIBLE);
-            }
-        });
-    }
-}
