--- app/src/main/java/com/example/android/alarmapp/MissionFragment.java
@@ -1,28 +1,0 @@
-package com.example.android.alarmapp;
-
-import android.app.AlertDialog;
-import android.app.Dialog;
-import android.content.DialogInterface;
-import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.v4.app.DialogFragment;
-
-public class MissionFragment extends DialogFragment {
-    public MissionFragment() {}
-
-    @NonNull
-    @Override
-    public Dialog onCreateDialog(Bundle savedInstanceState) {
-        AlertDialog.Builder builder = new AlertDialog.Builder(getActivity());
-
-        builder.setMessage(R.string.mission_string)
-                .setPositiveButton("OK", new DialogInterface.OnClickListener() {
-                    @Override
-                    public void onClick(DialogInterface dialog, int which) {
-                        // todo
-                    }
-                });
-
-        return builder.create();
-    }
-}
