--- app/src/main/res/layout/activity_add_alarm.xml
@@ -1,159 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<ScrollView
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-
-<LinearLayout
-    android:id="@+id/activity_add_alarm"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context="com.example.android.alarmapp.AddAlarmActivity"
-    android:orientation="vertical"
-    android:focusable="true"
-    android:focusableInTouchMode="true"
-    >
-
-    <TimePicker
-        android:id="@+id/tp_time_picker"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="center"
-        />
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginLeft="8dp"
-        android:layout_marginRight="8dp"
-        android:orientation="horizontal"
-        >
-        <ToggleButton
-            android:id="@+id/tgbtn_sun_repeat"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_weight="1"
-            android:textOn="@string/sunday_week"
-            android:textOff="@string/sunday_week"
-            />
-        <ToggleButton
-            android:id="@+id/tgbtn_mon_repeat"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_weight="1"
-            android:textOn="@string/monday_week"
-            android:textOff="@string/monday_week"
-            />
-        <ToggleButton
-            android:id="@+id/tgbtn_tue_repeat"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_weight="1"
-            android:textOn="@string/tuesday_week"
-            android:textOff="@string/tuesday_week"
-            />
-        <ToggleButton
-            android:id="@+id/tgbtn_wed_repeat"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_weight="1"
-            android:textOn="@string/wednesday_week"
-            android:textOff="@string/wednesday_week"
-            />
-        <ToggleButton
-            android:id="@+id/tgbtn_thr_repeat"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_weight="1"
-            android:textOn="@string/thursday_week"
-            android:textOff="@string/thursday_week"
-            />
-        <ToggleButton
-            android:id="@+id/tgbtn_fri_repeat"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_weight="1"
-            android:textOn="@string/friday_week"
-            android:textOff="@string/friday_week"
-            />
-        <ToggleButton
-            android:id="@+id/tgbtn_sat_repeat"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_weight="1"
-            android:textOn="@string/saturday_week"
-            android:textOff="@string/saturday_week"
-            />
-    </LinearLayout>
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="horizontal"
-        android:gravity="center_vertical"
-        android:layout_marginTop="12dp"
-        >
-
-        <Switch
-            android:id="@+id/swc_repeat"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginLeft="20dp"
-            android:text="@string/repeat_label"
-            android:textOn="@string/turn_on"
-            android:textOff="@string/turn_off"
-            />
-
-        <Switch
-            android:id="@+id/swc_vibrate"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginLeft="16dp"
-            android:text="@string/vibrate_label"
-            android:textOn="@string/turn_on"
-            android:textOff="@string/turn_off"
-            />
-
-        <Switch
-            android:id="@+id/swc_sound"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginLeft="16dp"
-            android:text="@string/sound_label"
-            android:textOn="@string/turn_on"
-            android:textOff="@string/turn_off"
-            />
-
-    </LinearLayout>
-
-    <EditText
-        android:id="@+id/et_memo"
-        android:inputType="textMultiLine"
-        android:scrollbars="vertical"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="16dp"
-        android:layout_marginLeft="8dp"
-        android:layout_marginRight="8dp"
-        android:hint="메모"
-        />
-
-    <Button
-        android:id="@+id/btn_save_alarm"
-        android:layout_width="120dp"
-        android:layout_height="wrap_content"
-        android:layout_gravity="center|bottom"
-        android:layout_marginTop="32dp"
-        android:layout_marginBottom="4dp"
-        android:text="저  장"
-        android:textSize="16dp"
-        android:textStyle="bold"
-        android:textColor="#FFFFFF"
-        android:background="@drawable/rounded_corner_btn"
-        />
-
-</LinearLayout>
-</ScrollView>
