--- app/src/main/res/layout/activity_detail_alarm.xml
@@ -1,121 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<ScrollView
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-
-<LinearLayout
-    android:id="@+id/activity_detail"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical"
-    tools:context="com.example.android.alarmapp.DetailAlarmActivity">
-
-
-    <TextView
-        android:id="@+id/tv_time_label"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="center_horizontal"
-        android:layout_marginTop="24dp"
-        android:textSize="36sp"
-        android:text="알람"
-        />
-
-    <ImageView
-        android:id="@+id/iv_alarm_icon"
-        android:layout_width="100dp"
-        android:layout_height="100dp"
-        android:layout_gravity="center_horizontal"
-        android:layout_marginTop="8dp"
-        android:background="@drawable/ic_alarm_on_blue_24dp"
-        />
-
-
-    <RelativeLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_gravity="center_horizontal"
-        android:layout_marginLeft="16dp"
-        >
-
-        <TextView
-            android:id="@+id/tv_am_pm"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="8dp"
-            android:layout_marginRight="8dp"
-            android:layout_alignBottom="@+id/tv_time"
-            android:layout_toLeftOf="@+id/tv_time"
-            android:textSize="32sp"
-            tools:text="AM"
-            />
-        <TextView
-            android:id="@+id/tv_time"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="8dp"
-            android:layout_centerInParent="true"
-            android:textSize="48sp"
-            tools:text="08:25"
-            />
-    </RelativeLayout>
-
-    <TextView
-        android:id="@+id/tv_location_label"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_margin="8dp"
-        android:layout_gravity="center_horizontal"
-        android:text="알람 설정 위치"
-        android:textSize="16sp"
-        />
-
-    <TextView
-        android:id="@+id/tv_location"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="center_horizontal"
-        android:layout_marginLeft="16dp"
-        android:layout_marginRight="16dp"
-        android:layout_marginBottom="8dp"
-        tools:text="jeju"
-        android:textSize="32sp"
-        />
-
-    <TextView
-        android:id="@+id/tv_memo_label"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="center_horizontal"
-        android:text="메모"
-        android:textSize="16sp"
-        />
-
-    <TextView
-        android:id="@+id/tv_memo"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginLeft="16dp"
-        android:layout_marginRight="16dp"
-        android:layout_marginTop="8dp"
-        android:layout_marginBottom="8dp"
-        tools:text="memo text memo text memo text memo text memo text memo text memo text "
-        android:textSize="16sp"
-        />
-    <Button
-        android:id="@+id/btn_alarm_stop"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_margin="16dp"
-        android:text="알람 중지"
-        android:textSize="36sp"
-        android:textColor="@android:color/white"
-        android:background="@android:color/holo_red_light"
-        />
-
-</LinearLayout>
-</ScrollView>
