--- app/src/main/res/layout/alarm_list_item.xml
@@ -1,160 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    xmlns:tools="http://schemas.android.com/tools">
-
-
-    <TextView
-        android:id="@+id/tv_am_pm"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentLeft="true"
-        android:layout_alignBottom="@+id/tv_time"
-        android:layout_marginLeft="@dimen/activity_horizontal_margin"
-        android:textSize="16sp"
-        tools:text="AM"
-        />
-
-    <TextView
-        android:id="@+id/tv_time"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_toRightOf="@+id/tv_am_pm"
-        android:layout_toEndOf="@+id/tv_am_pm"
-        android:layout_marginTop="8dp"
-        android:textSize="36sp"
-        tools:text="08:25"
-        />
-
-    <ToggleButton
-        android:id="@+id/tgbtn_vibrate"
-        android:layout_width="36dp"
-        android:layout_height="36dp"
-        android:layout_alignBottom="@+id/tv_time"
-        android:layout_toRightOf="@+id/tv_time"
-        android:layout_marginLeft="16dp"
-        android:background="@drawable/ic_vibration_on"
-        android:textOn=""
-        android:textOff=""
-        />
-
-    <ToggleButton
-        android:id="@+id/tgbtn_sound"
-        android:layout_width="36dp"
-        android:layout_height="36dp"
-        android:layout_alignBottom="@+id/tv_time"
-        android:layout_toRightOf="@+id/tgbtn_vibrate"
-        android:layout_marginLeft="8dp"
-        android:background="@drawable/ic_sound_on"
-        android:textOn=""
-        android:textOff=""
-        />
-
-    <!-- Weeks -->
-    <TextView
-        android:id="@+id/tv_sun"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentLeft="true"
-        android:layout_below="@+id/tv_time"
-        android:layout_marginLeft="@dimen/activity_horizontal_margin"
-        android:layout_marginTop="8dp"
-        android:textSize="12sp"
-        android:text="일"
-        />
-    <TextView
-        android:id="@+id/tv_mon"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_toRightOf="@+id/tv_sun"
-        android:layout_below="@+id/tv_time"
-        android:layout_marginLeft="8dp"
-        android:layout_marginTop="8dp"
-        android:textSize="12sp"
-        android:text="월"
-        />
-    <TextView
-        android:id="@+id/tv_tue"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_toRightOf="@+id/tv_mon"
-        android:layout_below="@+id/tv_time"
-        android:layout_marginLeft="8dp"
-        android:layout_marginTop="8dp"
-        android:textSize="12sp"
-        android:text="화"
-        />
-    <TextView
-        android:id="@+id/tv_wed"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_toRightOf="@+id/tv_tue"
-        android:layout_below="@+id/tv_time"
-        android:layout_marginLeft="8dp"
-        android:layout_marginTop="8dp"
-        android:textSize="12sp"
-        android:text="수"
-        />
-    <TextView
-        android:id="@+id/tv_thr"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_toRightOf="@+id/tv_wed"
-        android:layout_below="@+id/tv_time"
-        android:layout_marginLeft="8dp"
-        android:layout_marginTop="8dp"
-        android:textSize="12sp"
-        android:text="목"
-        />
-    <TextView
-        android:id="@+id/tv_fri"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_toRightOf="@+id/tv_thr"
-        android:layout_below="@+id/tv_time"
-        android:layout_marginLeft="8dp"
-        android:layout_marginTop="8dp"
-        android:textSize="12sp"
-        android:text="금"
-        />
-    <TextView
-        android:id="@+id/tv_sat"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_toRightOf="@+id/tv_fri"
-        android:layout_below="@+id/tv_time"
-        android:layout_marginLeft="8dp"
-        android:layout_marginTop="8dp"
-        android:textSize="12sp"
-        android:text="토"
-        />
-
-    <ImageButton
-        android:id="@+id/btn_edit_alarm"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_toLeftOf="@+id/btn_delete_alarm"
-        android:layout_alignBottom="@+id/tv_time"
-        android:layout_marginRight="8dp"
-        android:src="@drawable/ic_mode_edit_black_24dp"
-        />
-
-    <ImageButton
-        android:id="@+id/btn_delete_alarm"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignBottom="@+id/tv_time"
-        android:layout_alignParentRight="true"
-        android:layout_marginRight="16dp"
-        android:src="@drawable/ic_delete_black_24dp"
-        />
-
-    <View
-        android:layout_width="match_parent"
-        android:layout_height="1dp"
-        android:layout_alignParentBottom="true"
-        android:background="@android:color/darker_gray"
-        />
-
-</RelativeLayout>
