+++ app/src/main/res/layout/test_layout.xml
@@ -5,7 +5,7 @@
     android:id="@+id/activity_main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.example.android.alarmapp.DetailAlarmActivity">
+    tools:context="com.example.android.alarmapp.CameraActivity">
 
     <com.wonderkiln.camerakit.CameraView
         android:id="@+id/cameraView"
@@ -51,41 +51,14 @@
         >
 
         <android.support.design.widget.FloatingActionButton
-            android:id="@+id/btnToggleCamera"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            app:fabSize="normal"
-            android:src="@drawable/ic_cached_black_24dp"
-            android:tint="#ffffff"
-            android:layout_alignParentStart="true"
-            android:layout_alignParentLeft="true"
-            android:layout_marginLeft="32dp"
-            />
-
-        <android.support.design.widget.FloatingActionButton
             android:id="@+id/btnDetectObject"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             app:fabSize="normal"
             android:src="@drawable/ic_camera_alt_black_24dp"
             android:tint="#ffffff"
-            android:layout_centerInParent="true"
-            android:layout_centerVertical="true"
             />
-
-        <Button
-            android:id="@+id/btn_test_cancel"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="stop alarm"
-            android:layout_alignParentEnd="true"
-            android:layout_alignParentRight="true"
-            />
-
-
 
     </RelativeLayout>
-
-
 
 </FrameLayout>
